moved stream.h, state.h, and sourcepos.h into parsec/base/
