+++ include/cparsec3/parsec/base/sourcepos.h
@@ -1,8 +1,6 @@
 /* -*- coding: utf-8-unix -*- */
 #pragma once
 
-#include "../base/base.h"
-
 typedef struct SourcePos SourcePos;
 struct SourcePos {
   int line;
