+++ include/cparsec3/parsec/base/stream.h
@@ -1,8 +1,8 @@
 /* -*- coding: utf-8-unix -*- */
 #pragma once
 
-#include "../base/base.h"
-#include "State.h"
+#include "../../base/base.h"
+#include "state.h"
 
 #define Stream(S) TYPE_NAME(Stream, S)
 #define Token(S) TYPE_NAME(Token, S)
