+++ test/testit/src/cparsec3/stream/test_stream_string_takeN.c
@@ -1,5 +1,5 @@
 /* -*- coding: utf-8-unix -*- */
-#include <cparsec3/parsec/stream.h>
+#include <cparsec3/parsec/base/stream.h>
 
 #define S String
 #define T_GENERATOR FUNC_NAME(S, generator)
